Use uintptr_t in FreeRTOS_DNS.c as well

+++ FreeRTOS-Plus/Source/FreeRTOS-Plus-TCP/FreeRTOS_DNS.c
@@ -883,7 +883,7 @@ static const DNSMessage_t xDefaultPartDNSHeader =
 
 		/* Fill in the byte count, then move the pucStart pointer up to
 		the found byte position. */
-		*pucStart = ( uint8_t ) ( ( uint32_t ) pucByte - ( uint32_t ) pucStart );
+		*pucStart = ( uint8_t ) ( ( uintptr_t ) pucByte - ( uintptr_t ) pucStart );
 		( *pucStart )--;
 
 		pucStart = pucByte;
@@ -902,7 +902,7 @@ static const DNSMessage_t xDefaultPartDNSHeader =
 
 	/* Return the total size of the generated message, which is the space from
 	the last written byte to the beginning of the buffer. */
-	return ( ( uint32_t ) pucByte - ( uint32_t ) pucUDPPayloadBuffer + 1U ) + sizeof( DNSTail_t );
+	return ( ( uintptr_t ) pucByte - ( uintptr_t ) pucUDPPayloadBuffer + 1U ) + sizeof( DNSTail_t );
 }
 /*-----------------------------------------------------------*/
